logIn, logOut

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.koreaIT.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,14 +27,18 @@ public class UsrArticleController {
 	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요");
+		}
 		
 		if (Util.empty(title)) {
-			return ResultData.from("F-1", "제목을 입력해주세요");
+			return ResultData.from("F-2", "제목을 입력해주세요");
 		}
 		
 		if (Util.empty(body)) {
-			return ResultData.from("F-2", "내용을 입력해주세요");
+			return ResultData.from("F-3", "내용을 입력해주세요");
 		}
 		
 		articleService.writeArticle(title, body);
@@ -64,12 +70,16 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+			return ResultData.from("F-2", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -79,12 +89,16 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession httpSession, int id) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+			return ResultData.from("F-2", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.deleteArticle(id);
